refactor(spectral tools): spectrogram color files renamed
