+++ src/spectrogram/internal/macromagic.h
