+++ src/spectrogram/internal/roseuscolormaps.h
