+++ src/spectrogram/internal/spectrogramcolordefinitions.h
@@ -43,7 +43,7 @@ from there.  Audacity will look for a file called "Pause.png".
 // Note: No '#ifndef/#define' pair on this header file.
 // we want to include it multiple times in Theme.cpp.
 
-#include "MacroMagic.h"
+#include "macromagic.h"
 
 SET_THEME_FLAGS(resFlagNewLine);
