+++ src/spectrogram/internal/spectrogramcolorregister.h
