+++ src/spectrogram/spectrogramcolors.h
